requirements.txt
pybids>=0.9.*
